+++ modules/canvaskit/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2019-09-18
+
 ### Added
  - `SkCanvas.drawCircle()`, `SkCanvas.getSaveCount()`
  - `SkPath.offset()`, `SkPath.drawOval`